requirements.txt
opencv-python==4.3.0.36
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Session==0.3.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
Jinja2==2.11.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
random2==1.0.1
requests==2.23.0
requests-oauthlib==1.3.0
WTForms==2.2.1
